Typos, nit and more concise normative variants sections (#45)
* Typos, nit and more concise normative variants sections

I've gone through the normative variants sections and fixed typos, some style issues and tried to make it more concise. Some are for internal consistency or consistency with the existing PEPs, others are for clarity or for reducing non-normative aspects.

* Typos, nit and more concise normative variants sections

I've gone through the normative variants sections and fixed typos, some style issues and tried to make it more concise. Some are for internal consistency or consistency with the existing PEPs, others are for clarity or for reducing non-normative aspects.