Make the plugin API description more concise (#108)
* Wheel Variants: Make the plugin API description more concise

The plugin API documentation has been split into sections and made quite
verbose with the assumption that there will be more of it.  However,
we have been able to cut it down to one helper class, and a plugin
interface consisting of two attributes and two consistent functions.
Let's cut down on the sections, and describe the functions as a list
instead.

Signed-off-by: Michał Górny <mgorny@quansight.com>

* Wheel Variants: Clarify what "variant feature config" is for

Signed-off-by: Michał Górny <mgorny@quansight.com>

* Variant Wheels: remove `@runtime_checkable`

This is not really needed for documentation purposes, and unnecessarily
increases vertical length.

Signed-off-by: Michał Górny <mgorny@quansight.com>

* Wheel Variants: improve plugin API docstrings

Signed-off-by: Michał Górny <mgorny@quansight.com>

---------

Signed-off-by: Michał Górny <mgorny@quansight.com>